release(Root): Add initial codebase for nestjs project
This commit includes the initial codebase for the nestjs project. It provides a foundation for

future development and includes basic functionality such as user authentication and database

integration. The code has been organized and commented for clarity and ease of maintenance. Overall,

this commit lays the groundwork for a robust and scalable project